Moved flight functionality from ShipService to new FlightService.
