+++ service/src/main/java/restwars/service/flight/DetectedFlightDAO.java
@@ -1,4 +1,6 @@
-package restwars.service.ship;
+package restwars.service.flight;
+
+import restwars.service.ship.DetectedFlight;
 
 import java.util.List;
 import java.util.UUID;
