+++ service/src/main/java/restwars/service/flight/FlightDirection.java
@@ -1,4 +1,4 @@
-package restwars.service.ship;
+package restwars.service.flight;
 
 public enum FlightDirection {
     OUTWARD(0),
