+++ service/src/main/java/restwars/service/flight/FlightException.java
@@ -1,4 +1,4 @@
-package restwars.service.ship;
+package restwars.service.flight;
 
 import restwars.service.ServiceException;
