+++ service/src/main/java/restwars/service/flight/impl/flighthandler/AbstractFlightHandler.java
@@ -1,15 +1,21 @@
-package restwars.service.ship.impl.flighthandler;
+package restwars.service.flight.impl.flighthandler;
 
 import com.google.common.base.Preconditions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import restwars.mechanics.ShipMechanics;
 import restwars.service.event.EventService;
+import restwars.service.flight.DetectedFlightDAO;
+import restwars.service.flight.Flight;
+import restwars.service.flight.FlightDAO;
+import restwars.service.flight.FlightDirection;
 import restwars.service.infrastructure.RoundService;
 import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.planet.PlanetDAO;
 import restwars.service.resource.Resources;
-import restwars.service.ship.*;
+import restwars.service.ship.Hangar;
+import restwars.service.ship.HangarDAO;
+import restwars.service.ship.Ships;
 import restwars.service.ship.impl.ShipUtils;
 import restwars.util.MathExt;
