+++ service/src/main/java/restwars/service/flight/impl/flighthandler/TransportFlightHandler.java
@@ -1,17 +1,17 @@
-package restwars.service.ship.impl.flighthandler;
+package restwars.service.flight.impl.flighthandler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import restwars.mechanics.ShipMechanics;
 import restwars.service.event.EventService;
+import restwars.service.flight.DetectedFlightDAO;
+import restwars.service.flight.Flight;
+import restwars.service.flight.FlightDAO;
 import restwars.service.infrastructure.RoundService;
 import restwars.service.infrastructure.UUIDFactory;
 import restwars.service.planet.Planet;
 import restwars.service.planet.PlanetDAO;
 import restwars.service.resource.Resources;
-import restwars.service.ship.DetectedFlightDAO;
-import restwars.service.ship.Flight;
-import restwars.service.ship.FlightDAO;
 import restwars.service.ship.HangarDAO;
 
 import java.util.Optional;
