+++ service/src/main/java/restwars/service/telescope/impl/TelescopeServiceImpl.java
@@ -8,13 +8,14 @@ import restwars.service.building.Building;
 import restwars.service.building.BuildingDAO;
 import restwars.service.building.BuildingType;
 import restwars.service.event.EventService;
+import restwars.service.flight.*;
 import restwars.service.infrastructure.RandomNumberGenerator;
 import restwars.service.infrastructure.RoundService;
 import restwars.service.planet.Location;
 import restwars.service.planet.Planet;
 import restwars.service.planet.PlanetDAO;
 import restwars.service.player.Player;
-import restwars.service.ship.*;
+import restwars.service.ship.DetectedFlight;
 import restwars.service.telescope.PlanetWithOwner;
 import restwars.service.telescope.ScanException;
 import restwars.service.telescope.TelescopeService;
