+++ storage/src/main/java/restwars/storage/ship/JooqDetectedFlightDAO.java
@@ -5,11 +5,11 @@ import org.jooq.Record;
 import org.jooq.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import restwars.service.flight.DetectedFlightDAO;
+import restwars.service.flight.DetectedFlightWithSender;
+import restwars.service.flight.Flight;
 import restwars.service.player.Player;
 import restwars.service.ship.DetectedFlight;
-import restwars.service.ship.DetectedFlightDAO;
-import restwars.service.ship.DetectedFlightWithSender;
-import restwars.service.ship.Flight;
 import restwars.service.unitofwork.UnitOfWorkService;
 import restwars.storage.jooq.AbstractJooqDAO;
 import restwars.storage.mapper.DetectedFlightMapper;
