+++ storage/src/main/java/restwars/storage/ship/JooqFlightDAO.java
@@ -6,8 +6,13 @@ import org.jooq.Record;
 import org.jooq.Result;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import restwars.service.flight.Flight;
+import restwars.service.flight.FlightDAO;
+import restwars.service.flight.FlightType;
 import restwars.service.planet.Location;
-import restwars.service.ship.*;
+import restwars.service.ship.Ship;
+import restwars.service.ship.ShipType;
+import restwars.service.ship.Ships;
 import restwars.service.unitofwork.UnitOfWorkService;
 import restwars.storage.jooq.AbstractJooqDAO;
 import restwars.storage.mapper.FlightMapper;
